% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPPA_datasets.R
\docType{data}
\name{RPPA.20160128}
\alias{ACC.RPPA.20160128}
\alias{ACC.RPPA.20160128,BLCA.RPPA.20160128,BRCA.RPPA.20160128,CESC.RPPA.20160128,CHOL.RPPA.20160128,COADREAD.RPPA.20160128,DLBC.RPPA.20160128,ESCA.RPPA.20160128,GBMLGG.RPPA.20160128,HNSC.RPPA.20160128,KICH.RPPA.20160128,KIPAN.RPPA.20160128,KIRC.RPPA.20160128,KIRP.RPPA.20160128,LIHC.RPPA.20160128,LUAD.RPPA.20160128,LUSC.RPPA.20160128,MESO.RPPA.20160128,OV.RPPA.20160128,PAAD.RPPA.20160128,PCPG.RPPA.20160128,PRAD.RPPA.20160128,SARC.RPPA.20160128,SKCM.RPPA.20160128,STAD.RPPA.20160128,STES.RPPA.20160128,TGCT.RPPA.20160128,THCA.RPPA.20160128,THYM.RPPA.20160128,UCEC.RPPA.20160128,UCS.RPPA.20160128,UVM.RPPA.20160128}
\alias{BLCA.RPPA.20160128}
\alias{BRCA.RPPA.20160128}
\alias{CESC.RPPA.20160128}
\alias{CHOL.RPPA.20160128}
\alias{COADREAD.RPPA.20160128}
\alias{DLBC.RPPA.20160128}
\alias{ESCA.RPPA.20160128}
\alias{GBMLGG.RPPA.20160128}
\alias{HNSC.RPPA.20160128}
\alias{KICH.RPPA.20160128}
\alias{KIPAN.RPPA.20160128}
\alias{KIRC.RPPA.20160128}
\alias{KIRP.RPPA.20160128}
\alias{LIHC.RPPA.20160128}
\alias{LUAD.RPPA.20160128}
\alias{LUSC.RPPA.20160128}
\alias{MESO.RPPA.20160128}
\alias{OV.RPPA.20160128}
\alias{PAAD.RPPA.20160128}
\alias{PCPG.RPPA.20160128}
\alias{PRAD.RPPA.20160128}
\alias{RPPA.20160128}
\alias{SARC.RPPA.20160128}
\alias{SKCM.RPPA.20160128}
\alias{STAD.RPPA.20160128}
\alias{STES.RPPA.20160128}
\alias{TGCT.RPPA.20160128}
\alias{THCA.RPPA.20160128}
\alias{THYM.RPPA.20160128}
\alias{UCEC.RPPA.20160128}
\alias{UCS.RPPA.20160128}
\alias{UVM.RPPA.20160128}
\title{RPPA datasets from TCGA project from 2016-01-28 release date}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.RPPA.20160128

BLCA.RPPA.20160128

BRCA.RPPA.20160128

CESC.RPPA.20160128

CHOL.RPPA.20160128

COADREAD.RPPA.20160128

DLBC.RPPA.20160128

ESCA.RPPA.20160128

GBMLGG.RPPA.20160128

HNSC.RPPA.20160128

KICH.RPPA.20160128

KIPAN.RPPA.20160128

KIRC.RPPA.20160128

KIRP.RPPA.20160128

LIHC.RPPA.20160128

LUAD.RPPA.20160128

LUSC.RPPA.20160128

MESO.RPPA.20160128

OV.RPPA.20160128

PAAD.RPPA.20160128

PCPG.RPPA.20160128

PRAD.RPPA.20160128

SARC.RPPA.20160128

SKCM.RPPA.20160128

STAD.RPPA.20160128

STES.RPPA.20160128

TGCT.RPPA.20160128

THCA.RPPA.20160128

THYM.RPPA.20160128

UCEC.RPPA.20160128

UCS.RPPA.20160128

UVM.RPPA.20160128
}
\arguments{
\item{metadata}{A logical indicating whether load data into the workspace (default, \code{FALSE}) or to only display the object's metadata (\code{TRUE}). See examples.}
}
\description{
Package provides RPPA datasets from The Cancer Genome Atlas Project for all cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package} and contain snapshots for the release date: \code{2016-01-28} . Visit \pkg{RTCGA} site: \url{http://rtcga.github.io/RTCGA/}.
Use cases, examples and information about datasets in \pkg{RTCGA.RPPA.20160128}  can be found here: \code{browseVignettes("RTCGA")}. Link to the data format explanation is in the package DESCRIPTION file.
}
\details{
\code{browseVignettes("RTCGA")}
}
\examples{

\dontrun{
 ACC.RPPA.20160128(metadata = TRUE) 
 ACC.RPPA.20160128(metadata = FALSE) 
 ACC.RPPA.20160128 
}

}
\keyword{datasets}

